[NOTASK] Adding confirmation step before wiping up a folder

+++ scripts/init.sh
@@ -38,11 +38,15 @@ fi
 pushd "$ROOT_PATH"
 
 if [ -n "${CLEAN}" ]; then
-    echo "################################"
-    echo -e "\t Restart environment"
-    echo "################################"
-	run "rm -rf ${DATA_DIR}"
-    echo -e "################################ \n"
+	echo -e "You are about to wipe out ${DATA_DIR} "
+    read -p "Are you sure? [N/y]" -n 1 -r
+	echo    # (optional) move to a new line
+	if [[ $REPLY =~ ^[Yy]$ ]]; then
+	    echo -e "\t Restart environment"
+	    echo "################################"
+	    run "rm -rf ${DATA_DIR}"
+		echo -e "################################ \n"
+	fi
 fi
 
 if [ -n "${IS_DEBUG}" ]; then
